Add files via upload
(The data to database file). The system to accept and store sensor data in a SQLite database is established using the provided Python script. It specifies classes and methods for handling database queries and decoding incoming JSON data that includes motion sensor readings, temperature, and humidity. The script inserts the data into relevant tables in the database based on the incoming data and the context (such as the room or type of sensor). 
(The second file). The "Main.db" SQLite database is initialised using this code, along with other tables to hold various kinds of sensor data and related alerts. Each table has columns for the sensor ID, room, date, and associated readings or states, such as TemperatureData, HumidityData, MotionData, and LightingData. Furthermore, tables with names like TempWarning and MotionWarning are made to record cautions about motion and temperature, respectively. The script creates an organised environment for storing sensor data and warnings in the SQLite database "Main.db" by using the sqlite3 library to perform SQL instructions to define the table schema, shut the cursor, and end the database connection.